chatbot-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
psycopg2-binary
qdrant-client>=1.7.0
python-multipart==0.0.6
cryptography==41.0.8
requests==2.31.0
beautifulsoup4==4.12.2